[graphite] Enable contextOptions flag for readPixels during testing Dawn
During normal operation Graphite does not support synchronous
readPixels through the graphite::Image and graphite::Surface objects
because there's no access to a graphite::Context. However, this
operation is useful during testing, and it allows for a uniform
interface for both Ganesh and Graphite. The solution is to store a
reference to the Context on the current Recorder, and this is
indicated with the fStoreContextRefInRecorder ContextOptions flag.

Change-Id: I761ba03dd8043ea4b2fdeb1b39a11108a8684339
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/632557
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>