SkJpegCodec: Add HDRGM support
Add SkJpegGetHDRGMGainmapInfo, which will parse HDRGM XMP metadata to
populate an SkGainmapInfo.

Change the language in SkGainmapInfo to reflect the fact that the base
image can be HDR or SDR, and add a member variable indicating which
case is in play.

Make note of a bug where the fEpsilon members need to be changed from
scalar to vector (HDRGM allows vectors). For the moment, average the
epsilon values (to not break concurrent development).

Set the decoder to save segments with APP15 markers. If the call to
SkJpegGetHDRGMGainmapInfo succeeds, piece together the APP15 segments
to produce the SkData to return (copied, and as a stream).

Add a test for this format, too.

Bug: skia:14031
Change-Id: I81aeb02d2f56f13ada44983721c57ecd7aa8c83a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/631100
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>