Allow SK_TRIVIAL_ABI to be configured
Rename SK_SP_TRIVIAL_ABI to SK_TRIVIAL_ABI and allow it to be configured
by the build or by SkUserConfig.h. In the Skia build, default it to
being off with is_official_build, since users need to opt-in to an ABI
breaking change (which their application build must set and use
consistently).

Change-Id: I2c31c224b0f3fb2405f855d32faa161229402449
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633089
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>