[canvaskit] Control trivial_abi in compile_gm.sh
When trivial_abi is used it must be set and used consistently for the
entire build. `compile_gm.sh` breaks its build into two parts, the first
of which is a `gn` based build and the second of which is a manual build
of gms which links in the libraries created in the `gn` build. Ensure
that trivial_abi is set consistently in these two builds.

Change-Id: Ia7a432456dbccc5fe3cfc72d98e87935475b6202
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633957
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>