Add Analysis check for dynamically-uniform expressions.
We can generate simpler code if we know that an expression is
uniform across all lanes. If an expression consists of only literals,
uniforms, const variables, and pure function calls, we can safely
assume it will be the same on every lane. This is a very conservative
heuristic but is still enough to find a few wins.

Change-Id: If0d9efb725b93c71612a70a2f3531e735b804424
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633979
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>