Remove unnecessary use of SkOverrideDeviceMatrixProvider
We were using this in places where the matrix didn't represent the
SkDevice's local to device space transform. However, it still represents
the transformation from the shader's space to device space in these
cases and therefore it can be used for mip level and filter selection.

Add a GM for clip shader that verifies that matrix manipulation after
the clip shader is added to the canvas works correctly and that
clip image shader filter optimization doesn't break.

Bug: skia:13752
Change-Id: I0a756066b5b414caab576f468e5730717683031c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633178
Commit-Queue: Brian Salomon <bsalomon@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>