[sksl][wgsl] Correctly handle vector equality operator
In WGSL, the equality and inequality operators (==, !=) are
component-wise and result in a vector when applied to vector arguments.
In SkSL the result needs to be converted to a boolean when the
expression needs to evaluate to a boolean.

This is now done for vectors using the all() and any() intrinsics.

Bug: skia:13092
Change-Id: I3377e06ad0703cad38fb11447001a500ebe5ed54
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/631168
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>