Seed coords earlier in shader traversal for SkRasterPipeline.
This enables shader coord manipulation. Previously coords were seeded in and transformed in leaf SkShaders.

A new type, SkShaderBase::MatrixRec, is added. It tracks whether device coords were already seeded and a "pending" matrix that transforms them to the shader's coordinate space. SkShader subclasses are responsible for "applying" the MatrixRec before adding anything to the raster pipeline that requires the coordinates be seeded and in the right space.

The MatrixRec also tracks whether the total matrix between device and shader space is known. This allows SkShaders to avoid optimizations that depend on that knowledge. The matrix becomes unknown in two cases at present:

1) SkTransformShader. This shader inserts a matrix that can be updated after the SkRasterPipeline is created, between executions.

2) SkRTShader. The SkSL code may manipulate the coords directly in ways that aren't captured by MatrixRec.

Bug: skia:13752
Change-Id: I5e2ba2fcddad22c5231febab49d50c140425a2b6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/628742
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>