Change branch-if-all-lanes-equal to branch-if-no-active-lanes-equal.
This op was originally made for dynamically-uniform comparisons,
but by changing its definition slightly, it will also be good for
building switch-case blocks. Specifically:

- Dead lanes are masked off. This didn't matter for a dynamically-
uniform expression, but matters a great deal for pretty much
everything else.
- "Branch-if-not-equal" is just as useful as "branch-if-equal" when
distinguishing between a 'true' or 'false,' but the not-equal op
will let us branch past unused switch-cases in a single op. (If the
case is live, we will still need to set the condition mask
appropriately so that we only affect the proper lanes.)

Change-Id: Iae78c6a0aee27ceb99808472a269c0ef323f38b6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633975
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>