Simplify for-loops with straight-line control flow.
If we know that for-loop won't escape early due to a `continue`,
`break` or `return`, and the loop conforms to ES2 structure, we
are guaranteed to run the full number of iterations across all
executing lanes. In this case, we don't need to use a loop mask.

Change-Id: I75fd1aa1c929219230a104ed19b4baa62e8e90fc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/634458
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>