Add mechanism for recycling temp slots.
Previously, every time we used a value slot for something ephemeral
(e.g. a continue-mask in a loop), the slot was lost forever. Now,
it is added to a recycle-list and can be reclaimed.

I think I will need a temporary value slot to implement non-constant
array indexing, and it would be pretty wasteful if every array index
burned a slot forever.

Change-Id: I1a31e7cea8b2c437c9d00d47c709e980694df401
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/635179
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>