Replace Pathops in Geometry::ChopMonoCubic implementation
This uses code in base instead, breaking the connection between
SkPath and Pathops.

This extracts one more function into SkBezierCurves.h, one
that converts between the control points and the polynomial.
It also adds a test comparing that functionality to the
Pathops one. We might be able to, in a follow-up CL, replace
SkDCubic::Coefficients with SkBezierCubic::ConvertToPolynomial.

Change-Id: I65e2b692c63a8c647ac48552008424ef1ea5d737
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633458
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>