[pdf] SkPDFGraphicStackState::Entry::fColor default
Make the default value for SkPDFGraphicStackState::Entry::fColor a very
invalid color instead of opaque black in order to better match the
actual defaults in PDF.

Currently if an SkPDFDevice is created and the first thing done with it
is to draw an opaque black path then the opaque black will not be
written to the PDF since it is currently assumed that the default will
be opaque black. However, because no color space is currently defined
for the fill and stroke colors and PDF otherwise has no default, the
path will instead be filled with an indeterminate opaque color (in the
case of poppler this will be cyan).

Change-Id: I9f6a46c42fe8721d6a6f3dddc5ac986207a64bcd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/635602
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>