SkJpegCodec: Add SkJpegSourceMgr (reland)
The current skjpeg_source_mgr (in SkJpegUtility.cpp/h) is effectively
two implementations (buffered vs in-memory) controlled via if
statements.

Turn these into two implementations of a pure virtual interface
SkJpegSourceMgr. The reason for this it that there will be a third
interface added for buffered non-seekable sources (which will
run an SkJpegSegmentScanner on the data as it is read).

Leave the pre-existing skjpeg_source_mgr in place because it may
be used in other places. It will be removed separately.

This was reverted because of Chromium roll issues related to adding
the new SkJpegSourceMgr.cpp/h files.

Bug: skia:14031
Change-Id: If7ccf386dd17de8eaf468bb8e81714402173c146
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633465
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>