Add atan2 support to SkRasterPipeline builder/codegen.
Since atan2 generates a lot of code, I decided not to have dedicated
1-4 slot versions of the op; we only have an n-way op, which can
handle any number of elements. This is the first time we've done
this with a binary op, so it required a bit of restructuring, but
nothing too spooky.

Change-Id: I06c93277dc803a6b83a8932009c2a519fdcf8546
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/635717
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>