Remove skvx usage from PipelineData and UniformManager
As a follow-up to https://skia-review.googlesource.com/c/skia/+/635796
the skvx types are not needed for speed/computation in
PipelineData nor UniformManager. This replaces them with
SkV2 and SkV4 (defined in SkM44.h).

While I was looking, I added the include for SkVx.h to all
the places in Graphite that use it to make the dependency
more clear. It seems SkVx.h is used in a few headers
(could lead to increased compile times); all those modified
here and geom/SkRect.h. We may want to move some of those
uses out of a header file, if appropriate.

This also adds const-ness to the SkPoint variant, as that
seems appropriate.

Change-Id: I3f88db4cd98e0f9eabc878a5b50612a446a66483
Bug: skia:13052
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/636676
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>