Add SK_CODEC_DECODES_JPEG guards
The file SkJpegCodec.cpp has a SK_CODEC_DECODES_JPEG that guards the
entire file. This was added in:
https://skia-review.googlesource.com/c/skia/+/506356

Add this guard to SkJpegSourceMgr.cpp, because it is included in the
build in the same places as SkJpegCodec.cpp. Also add this guard to
SkJpegSegmentScan.cpp, because soon it will have this property, too.

Bug: skia:14031
Change-Id: I8d84df0a68345fc602ee00881b4439a51d21e9b6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/637177
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>