Add SkJpegUnseekableSourceMgr
Add SkJpegUnseekableSourceMgr as a new sub-class of SkJpegSourcMgr.
This is for use with streams that cannot be seeked and do not
present and end-of-file (e.g, loading an image from a pipe).

This source manager presents data to a SkJpegSegmentScanner to be
scanned before feeding it to the decoder. It also tracks how many
bytes have been read since the start of the image.

Using these pieces, it can provide the marker offsets needed by the
various gainmap formats (e.g, EndOfImage for JpegR or APPN for MPF).

It can also implement getSubsetStream given an offset relative to the
start of the image.

Add lots of tests that create streams off all of the different
capability levels. Also add sub-tests where we skip various parts
of the internal image (e.g, skipping the middle image in an MPF image).

Bug: skia:14031
Change-Id: Ibd9f08c5cb6bcfb08ff5c74a7875bcda66909345
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/633463
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>