Fix type handling for matrix multiplication.
This code assumed that `type` would be equivalent to `left.type()`.
This is generally true, but is wrong for `scalar x matrix`.
(Interestingly, this error is only reached via the SkRuntimeEffect
usage of SkRP. I'm not yet sure why that is the case.)

Change-Id: I7be0a1d5beba01f262299fcbf06aa8e46c029e0d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/637504
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>