Fix handling of operator | and ||.
A copy-paste error led to || actually being implemented as bitwise-
and.

This was not discovered sooner because an operator|| with side
effects on the right will not take this path, and because operator|
is unsupported in ES2 and so is not heavily tested. However, our
SkSLTest code happens to run tests with optimization on and off, and
the TernaryFolding test managed to trigger the issue here:

https://osscs.corp.google.com/skia/skia/+/main:resources/sksl/folding/TernaryFolding.rts;l=18-19;drc=ad5437dbab33d2a5cf568e0132b0b593d1e5202c

With optimizations off, we would trigger this miscompilation.
(With optimizations on, this line would be eliminated since it is
effectively a no-op.)

Change-Id: I00ba3c7e130ef236769c0ba530fd0fb07aeaa20c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/637506
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>