Set the paint color in SkColorFilterBase::onFilterColor4f.
Previously, `onFilterColor4f` left the color on the paint black.
In SkRP, this caused the evaluation of a null child shader to
return black instead of the paint color, because
`RuntimeEffectRPCallbacks::appendShader` returns the paint color
when evaluating a null shader.

(This is still not quite perfect; if you have a paint which has
a paint color, a meaningful shader, _and_ a color filter which
evaluates a null shader, the null-shader evaluation should ideally
return the paint color with shading applied, not the original
paint color. However I don't think we have any unit tests which
exercise this case, and it's certainly not something we expect
any clients are relying on.)

Change-Id: If52afddc90d1ee5553a05751022c7999c5d68ba1
Bug: skia:14092
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/637696
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>