Implement linear mipmap filtering in Raster Pipeline
Includes some naming changes to align things with SkVM (and for better
clarity overall). The big change is stuffing everything that's specific
to a level (pixmap, matrix, and context pointers) into a helper struct,
and plubming that through the lambdas. This lets us reuse that code to
generate two sets of gather/filter stages.

Then, we just use a set of stages to save off the original coords, do a
switcheroo between levels, and a final lerp to get the real result.

Bug: skia:13678
Change-Id: I6b3cf124324f08022eecaf309d147bc657df8227
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/638028
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>