Simplify division by a constant.
An expression like `x / 4` is now rewritten as `x * 0.25` where
it's safe to do so. We hope that most GPUs are smart enough to do
this automatically, but in the software backend, this sort of
optimization is up to us. This was implemented in the constant
folder because that seemed easiest; I don't expect it to actually
improve any GPU results, but it's harmless and it does shake out
slightly simpler code in a few cases.

Change-Id: Id37011d6a94d1f213052fd551428a6ea08835520
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/637896
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>