SkJpegMultiPicture: Fix segment identification bug
JPEG Multi Picture Format (MPF) has an APP2 segment that contains the
offsets and sizes of the auxiliary images. The offsets are relative
to that APP2 segment.

We extract the MPF segment parameters using libjpeg, but libjpeg will
not tell us the offset into the file of the APP2 segment, so we have
no way of knowing how to find the auxiliary images.

Meanwhile, SkJpegSourceMgr's SkJpegSegmentScanner has the opposite
feature set -- it knows the offset into the file of the MPF segment,
but it doesn't save the parameters themselves.

In order to get both the MPF segment offset and the MPF segment
parameters, walk through libjpeg's APP2 marker_list while also walking
through the SkJpegSegmentScanner's APP2 segment list.

Prior to this change we just assumed that the first APP2 segment was
the MPF segment. This is not a safe assumption, because APP2 is also
used by ICC profiles.

A better solution to this problem may be to add the ability to query segment parameters to SkJpegSegmentScanner. This is quite do-able for
seekable streams, but gets complicated for unseekable streams.

Bug: skia:14031
Change-Id: Ia24104f616207398a51f0f40e6e0f344bde919ee
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/638796
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>