Use path and drawable API correctly
The bool returned from setPath and setDrawable does
not mean if they exist. It means if the code needs to
update the bytes used. Set them, and then check if they
exist using path() and drawable().

This should fix the problems with gold looking ugly
for the glremoteslug config.

Change-Id: Ia04c30c9cb31d0612553e51a615ce0c58c104c3a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/639616
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>