Build runtime files
When doing compile-only builds, we did not have Bazel generate
the metadata files that would be necessary to actually run the
executables, as those were thought to be unnecessary.

However, if a rule needs to use something like a py_binary [1]
to generate files, it appears those manifests *are* necessary.

This should unblock the VK-deps roller [2]

[1] https://github.com/KhronosGroup/SPIRV-Tools/blob/0ce2bc4ce0ac768b86047a85f31ccb5ea3263f31/BUILD.bazel#L101
[2] https://skia-review.googlesource.com/c/skia/+/639366

Change-Id: I439c4527f36ce9df8d0ac45e4339f79c647c8662
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/639641
Reviewed-by: Michael Ludwig <michaelludwig@google.com>