Fix microhttpd on Xcode 14.2.
It appears that microhttpd is using APIs like lseek and close
without #including <unistd.h> in a few places. We are not on
the newest microhttpd version, but I'm not sure if this issue is
actually fixed in the newest versions anyway.

Change-Id: I470e797667c3f490aa462a4bbcd5ec7587269e2e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/641057
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Herb Derby <herb@google.com>