[android-next] Remove CanvasKit/PathKit jobs from CQ
CanvasKit and PathKit jobs are currently broken on this old
branch due to the depot_tools ninja change. In general, we don't
want to support building the *Kits on any branch other than
the primary one, so we can just remove all related jobs.

While we are fixing up jobs, this includes
https://skia-review.googlesource.com/c/skia/+/631520
to fix the IntelIris540 jobs on the CQ.

Note for myself: I created this branch with
git new-branch --upstream=remotes/origin/android/next-release fix-canvaskit

Change-Id: Ia0f5fc571bfc81b73271da5981bf278e3ec54800
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/641476
Reviewed-by: Leon Scroggins <scroggo@google.com>