[graphite] Introduce ComputeStep interface
ComputeStep defines an interface to define a compute program and its
data flow within a broader rendering framework. It is the compute
pipeline equivalent of RenderStep and it is intended to be compatible
with Renderers managed by a RendererProvider. However, it is also meant
to be flexible and can be used as a building block for compute kernels
that apply image effects.

Subsequent CLs will incorporate ComputeStep into ComputePipeline,
ComputePass, and ComputePassTask creation.

Bug: b/240604614
Change-Id: I119afac58fc16e4b415314962bda079d2a41b6ca
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/641699
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>