[graphite] Use ComputeStep in compute pipeline creation and indexing
Previously, a compute pipeline was created using a shader snippet
assigned directly to a ComputePipelineDesc. The unique resource key for
a compute pipeline was created by hashing the SkSL string and the name
assigned to the shader.

This is now done using a ComputeStep. The shader module is assembled and
compiled in the backend specific ResourceProvider and the resource key
is based on the ComputeStep's unique ID.

This aligns compute with GraphicsPipeline creation and indexing which
follows a similar logic that involves a RenderStep.

Bug: b/240604614
Change-Id: I0265e3e3fbe2f22631b101a437f54fc207b80a97
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/641700
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>