[graphite-dawn] Correctly set the binding size of uniform buffers
Currently the binding size of uniform buffers are set to whole buffer
when creating a bind group. However, there are cases when a big uniform
buffer is created and only a subset of it is used as uniforms. This
could lead to validation error because max uniform buffer's binding size
is 64 KB and it's possible the buffer's total size is larger than that.

Fix by clamping the binding size of uniform buffers to 64KB. If the
binding size is less than 64 KB use kWholeBuffer constant.

Bug: b/268671347
Change-Id: I8eb8eb96daeb8e9128d495e184a197a15b6b0d0f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/641676
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>