Move SkRandom and SkFixed into base
SkFixed requires a shim for Android [1].

Client CL:
 - https://chromium-review.googlesource.com/c/chromium/src/+/4242158

SkFixed.h was moved to include/private/base because it is used
elsewhere in include/. SkRandom was not, so it is now in src/base.

See SkFixed.h, SkRandom.h, and the Bazel files for actual changes,
everything else was just find and replace.

The fiddles (docs/examples/*) were changed to use <random> because
we only want those to demonstrate Skia's public APIs.

The shim for SkTemplates.h no longer appears used.

[1] https://android.googlesource.com/platform/frameworks/base/+/refs/heads/master/libs/hwui/SkiaInterpolator.cpp#20

Change-Id: Ie4fddd33ffb4141c1f0dc051712407d0da64aa4b
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/641776
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>