Disable test SkSLIntrinsicMixFloatES3_GPU on Win + RTX3060.
This is currently red on the tree.

I reused an existing RTX3060 block for this and added a Windows OS
check. When I added the block originally, we didn't have any
non-Windows OSes on the tree with an RTX3060; it was always
intended to be a Windows-specific disable.

Change-Id: Ic7acd0290c25872a6a3f9bbf6b3e8db174d9bcb3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/642277
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>