Reorder the inputs to RP mix() stage.
By itself, this change doesn't affect anything other than the order
that we write arguments to the stack. For the mix intrinsic, this
order is irrelevant. However, putting the interpolation point before
the interpolants will allow us to rewrite `test ? x : y` as
`mix(test, y, x)` and preserves the property that the ternary test is
evaluated before the evaluation of x or y. (This also requires that
x and y have no side-effects.) This will be expanded upon in a
followup.

Change-Id: Idf0f1fadc41ce1ca266e3375fe3ada3f938b00a6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/642058
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>